Testing for DEX files w/ cases
