+++ src/test/resources/dex/894/dex.sample.4010.case.txt
@@ -2,14 +2,10 @@ DXS*8050880508*DX*004010UCS*113*9254850000
 ST*894*0101
 G82*D*001701001701*051957769*5*123456789*073002*20180416
 LS*0100
-G83*1*4*EA*007189983548****2.13*1*1/2 GAL TEST 1
-G22*Y*563
-G72*97*02***-0.35
-G83*2*4*EA*007189903720****2.25*1*1/2 GAL TEST 2
-G22*Y*548
+G83*1*2*CA*007800001180***007800001180*14*2*12z12P 7Up***
 LE*0100
 G84*8*4164
 G86*0840
 G85*8263
-SE*13*0101
+SE*9*0101
 DXE*113*1
