+++ src/test/resources/dex/894/dex.sample.5010.case.txt
@@ -1,0 +1,11 @@
+DXS*8050880508*DX*005010UCS*113*9254850000
+ST*894*0101
+G82*D*001701001701*051957769*5*123456789*073002*20180416
+LS*0100
+G83*1*1*CA**UK*00078000011807**14*2*12z12P 7Up***1
+LE*0100
+G84*8*4164
+G86*0840
+G85*8263
+SE*9*0101
+DXE*113*1
