--- src/test/resources/dex/894/dex.sample.cases.txt
@@ -1,96 +1,0 @@
-DXS*6150430000*DX*004010UCS*1*9254850000*P
-ST*894*0001
-G82*D*8327063806*051957769*004615*182737015*PL1124*20171116
-LS*0100
-G83*1*2*CA*007800001180***007800001180*14*2*12z12P 7Up***
-G72*497*02***-5.74
-G83*2*1*CA*007800001181***007800001181*14*2*12z12P Dt 7Up***
-G72*497*02***-5.74
-G83*3*3*CA*007800001182***007800001182*14*2*12z12P Cherry 7UP***
-G72*497*02***-5.74
-G83*4*3*CA*007800001188***007800001188*14*2*12z12P Dt Cherry 7Up***
-G72*497*02***-5.74
-G83*5*2*CA*007800004116***007800004116*14*2*12z12P RC***
-G72*497*02***-5.74
-G83*6*1*CA*007181703329***007181703329*14*2*12z12P BigRed***
-G72*497*02***-5.74
-G83*7*2*CA*007800005216***007800005216*14*2*12z12P AW RtBeer***
-G72*497*02***-5.74
-G83*8*6*CA*007800005316***007800005316*14*2*12z12p Dt AW RtBeer***
-G72*497*02***-5.74
-G83*9*1*CA*007800005416***007800005416*14*2*12z12P AW Cream Soda***
-G72*497*02***-5.74
-G83*10*1*CA*007800015216***007800015216*14*2*12z12P CD GingerAle***
-G72*497*02***-5.74
-G83*11*1*CA*007800014816***007800014816*14*2*12z12p Dt CD GingAle***
-G72*497*02***-5.74
-G83*12*2*CA*007800000034***007800000034*16*8*2l8 7Up***
-G72*497*02***-8.48
-G83*13*2*CA*007800000075***007800000075*16*8*2L8 Dt 7Up***
-G72*497*02***-8.48
-G83*14*1*CA*007800000531***007800000531*16*8*2L8 Cherry 7Up***
-G72*497*02***-8.48
-G83*15*1*CA*007800000337***007800000337*16*8*2L8 RC***
-G72*497*02***-8.48
-G83*16*1*CA*007800012246***007800012246*16*8*2L8 Dt SK Orange***
-G72*497*02***-8.48
-G83*17*2*CA*007800005246***007800005246*16*8*2l8 AW RtBeer***
-G72*497*02***-8.48
-G83*18*1*CA*007800001646***007800001646*16*8*2L8 Squirt***
-G72*497*02***-8.48
-G83*19*1*CA*007800015246***007800015246*16*8*2l8 CD GingerAle***
-G72*497*02***-8.48
-G83*20*1*CA*007800014846***007800014846*16*8*2L8 Dt CD GingerAle***
-G72*497*02***-8.48
-G83*21*1*CA*007800001219***007800001219*24*4*0.5L6P 7Up***
-G72*497*02***-14.35
-G83*22*1*CA*007800001223***007800001223*24*4*0.5L6P Dt 7Up***
-G72*497*02***-14.35
-G83*23*1*CA*007800005242***007800005242*24*4*0.5l6p AW RtBeer***
-G72*497*02***-14.35
-G83*24*1*CA*007800005342***007800005342*24*4*0.5L6P Dt AW RtBeer***
-G72*497*02***-14.35
-G83*25*1*CA*007800000117***007800000117*24*4*0.5L6p CD GAle***
-G72*497*02***-14.35
-G83*26*1*CA*084288509726***084288509726*24*12*16z12 Xy Fruit Punch***
-G72*497*02***-12
-G83*27*1*CA*084288509863***084288509863*24*12*16z12 XyBluPom***
-G72*497*02***-12
-G83*28*1*CA*018624900004***018624900004*24*12*16z12 XyCranRazz***
-G72*497*02***-12
-G83*29*1*CA*084288509701***084288509701*24*12*16z12 XyMelon Mayhem***
-G72*497*02***-12
-G83*30*1*CA*085817600205***085817600205*30*12*16Z12 BodyArmOragMgo***
-G72*497*02***-17.8
-G83*31*1*CA*085817600206***085817600206*30*12*16z12 BodyArmStrwBan***
-G72*497*02***-17.8
-G83*32*1*CA*007800000197***007800000197*14*4*7.5z6p A&W Root Beer***
-G72*497*02***-5.08
-LE*0100
-G84*47*40979
-G86*1310B1BF
-G85*4DFD
-SE*72*0001
-ST*894*0002
-G82*D*8327063807*051957769*004615*182737015*PL1124*20171116
-LS*0100
-G83*1*1*CA*063256500009***063256500009*24*4*0.5l6p FIJI***
-G72*497*02***-3.4
-LE*0100
-G84*1*2060
-G86*1310B1BF
-G85*2217
-SE*10*0002
-ST*894*0003
-G82*D*8327063808*051957769*004615*182737015*PL1124*20171116
-LS*0100
-G83*1*1*CA*089940700295***089940700295*26*12*14.5z12 Neuro Daily***
-G72*497*02***-8.4
-G83*2*1*CA*089940700301***089940700301*26*12*14.5z12 Nro BldOrgPa***
-G72*497*02***-8.4
-LE*0100
-G84*2*3520
-G86*1310B1BF
-G85*41E5
-SE*12*0003
-DXE*1*3
